Improve file and method naming
