+++ source/cli.js
@@ -6,7 +6,7 @@ const logSymbols = require('log-symbols');
 const meow = require('meow');
 const updateNotifier = require('update-notifier');
 const hasYarn = require('has-yarn');
-const {isPackageNameAvailable} = require('./npm-util');
+const {isPackageNameAvailable} = require('./npm/util');
 const version = require('./version');
 const util = require('./util');
 const ui = require('./ui');
