+++ source/version.js
@@ -3,11 +3,11 @@ const semver = require('semver');
 
 class Version {
 	constructor(version) {
-		this.version = version;
+		this.version = semver.coerce(version);
 	}
 
 	isPrerelease() {
-		return module.exports.PRERELEASE_VERSIONS.includes(this.version) || Boolean(semver.prerelease(this.version));
+		return Boolean(semver.prerelease(this.version));
 	}
 
 	satisfies(range) {
@@ -45,6 +45,8 @@ module.exports = version => new Version(version);
 
 module.exports.SEMVER_INCREMENTS = ['patch', 'minor', 'major', 'prepatch', 'preminor', 'premajor', 'prerelease'];
 module.exports.PRERELEASE_VERSIONS = ['prepatch', 'preminor', 'premajor', 'prerelease'];
+
+module.exports.isPrereleaseOrIncrement = input => module.exports(input).isPrerelease() || module.exports.PRERELEASE_VERSIONS.includes(input);
 
 const isValidVersion = input => Boolean(semver.valid(input));
